fix(log): match the Swift tauri_log level type with Rust
Rust declares `tauri_log(level: u8, ...)` (plugins/log/src/lib.rs) but the
Swift side took `level: Int` (plugins/log/ios/Sources/LogPlugin.swift). The
C ABI does not guarantee the upper bits are zeroed, so the switch could
fall through to `default` and drop the record. The Swift parameter is now
UInt8. Unverified: Swift not compiled.

+++ plugins/log/ios/Sources/LogPlugin.swift
@@ -8,7 +8,8 @@ import UIKit
 import os.log
 
 @_cdecl("tauri_log")
-func log(level: Int, message: NSString) {
+// `level` must match the `u8` declared on the Rust side of this FFI function.
+func log(level: UInt8, message: NSString) {
   switch level {
   case 1: Logger.debug(message as String)
   case 2: Logger.info(message as String)
